818: update language for acs tooltips (remove puma references) and indicators methodology
