+++ app/templates/about.hbs
@@ -15,7 +15,7 @@
     <div class="grid-x grid-padding-x">
       <div class="cell xlarge-7">
 
-        <p>Indicators were carefully selected to present a snapshot across a broad range of conditions related to city planning. The {{fa-icon "info-circle"}} icon reveals an indicator’s source. Indicators using data from the American Community Survey (ACS) are calculated at the Public Use Microdata Area (PUMA) level. PUMAs are geographic approximations of community districts, but their boundaries differ slightly. <a href="http://www1.nyc.gov/assets/planning/download/pdf/data-maps/nyc-population/census2010/puma_cd_map.pdf" target="_blank">Use this map</a> to compare PUMA and community district boundaries. Some community districts share a PUMA with another district, and will therefore display identical values for ACS indicators. These paired districts include: Manhattan 1 and 2; Manhattan 4 and 5; Bronx 1 and 2; and Bronx 3 and 6. ACS population estimates are not available for districts that share a PUMA because they cannot be reliably disaggregated to each individual district.</p>
+        <p>Indicators were carefully selected to present a snapshot across a broad range of conditions related to city planning. The [i] icon reveals an indicator’s source.  Most indicators rely on data from the American Community Survey (ACS) or Decennial Census. Community districts are NYC-specific geographic bounds, so the Census Bureau does not report data at the community district level.  Instead, we need to compute community district estimates by approximating them from smaller census geographies.  To do this, we aggregate data from blocks and block-groups based on a geographic lookup relating each census block and tract to the community district it falls within. Using this lookup, we aggregate count data, then recompute rate estimates and margins of error. This methodology is used to compute the intercensal total number of persons within a community district, which serves as the denominator for some non-ACS rate indicators (such as “crimes per 1000”).</p>
 
       </div>
       <div class="cell xlarge-5">
